Improve testing with ENV files
Prior to this, we had a single env file in the repo itself that the tests were running against. Unfortunately this means that a single file needs to cover all our use cases. It also introduces a mystery guest into our tests.

This gives us a set of functions that allow us to create a new ENV file with the given input. It will use a random file name so we don't step on our own toes.

As we move forward, I'm guessing that we'll extract this logic into another module to be used for other tests but I think this is a good start.